financial_document_processor/requirements.txt
# PDF Extraction
unstructured>=0.10.0
camelot-py>=0.10.1
pdfplumber>=0.9.0
opencv-python>=4.7.0
pdf2image>=1.16.3
pytesseract>=0.3.10
poppler-utils>=0.1.0

# Database
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
alembic>=1.11.0

# AI and NLP
llama-index>=0.8.0
langchain>=0.0.267
openai>=1.0.0
google-generativeai>=0.3.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
